requirements.txt
aiodns==1.1.1
aiohttp==1.1.1
aiohttp-jinja2==0.12.0
aiopg==0.11.0
appnope==0.1.0
async-timeout==1.1.0
cchardet==1.1.0
chardet==2.3.0
decorator==4.0.10
Django==1.10.5
django-extensions==1.7.5
django-jquery==3.1.0
djangorestframework==3.5.3
flake8==3.0.4
gunicorn==19.6.0
ipython==5.1.0
ipython-genutils==0.1.0
Jinja2==2.8
MarkupSafe==0.23
marshmallow==2.10.3
mccabe==0.5.2
multidict==2.1.2
passlib==1.6.5
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.7
psycopg2==2.6.2
ptyprocess==0.5.1
py==1.4.31
pycares==2.1.1
pycodestyle==2.0.0
pyflakes==1.2.3
Pygments==2.1.3
PyJWT==1.4.2
pytest==3.0.2
pytest-aiohttp==0.1.3
python-dateutil==2.5.3
PyYAML==3.12
requests==2.11.1
simplegeneric==0.8.1
simplejson==3.8.2
six==1.10.0
SQLAlchemy==1.0.15
stripe==1.38.0
traitlets==4.3.1
wcwidth==0.1.7
Werkzeug==0.11.15
yarl==0.5.3
